fixed force-based contact detection cleaned up tests
