+++ src/roman/rq/__init__.py
@@ -1,3 +1,4 @@
 from .hand import *
 from .connection import *
+from .sim_connection import *
 from .controllers import *
