+++ src/roman/ur/scripts/constants.py
@@ -70,8 +70,8 @@ UR_TOOL_POSITION_TOLERANCE = 0.001 # m
 UR_TOOL_ROTATION_TOLERANCE = 0.01 # rad
 UR_DEADMAN_SWITCH_LIMIT = 0.1 # seconds
 
-UR_DEFAULT_FORCE_LOW_BOUND = [-20.0, -20.0, -20.0, -2.0, -2.0, -2.0] 
-UR_DEFAULT_FORCE_HI_BOUND = [20.0, 20.0, 20.0, 2.0, 2.0, 2.0]
+UR_DEFAULT_FORCE_LOW_BOUND = [-10.0, -10.0, -10.0, -1, -1, -1] 
+UR_DEFAULT_FORCE_HI_BOUND = [10.0, 10.0, 10.0, 1, 1, 1]
 UR_DEFAULT_ACCELERATION = 0.1 # rad/s2
 UR_FAST_STOP_ACCELERATION = 2.0 # rad/s2
 UR_DEFAULT_MAX_SPEED = 0.1  # rad/s
