+++ src/roman/ur/scripts/urlib.script
@@ -113,7 +113,11 @@ def ur_force_limit_exceeded(low_bound, high_bound):
         (high_bound[3] < ft[3]) or 
         (high_bound[4] < ft[4]) or 
         (high_bound[5] < ft[5])
-
+    if exceeded:
+        textmsg("force limit exceeded at: ", ft)
+        textmsg("low bound: ", low_bound)
+        textmsg("high bound: ", high_bound)
+    end
     return exceeded
 end
