+++ src/test/arm_functional_test.py
@@ -1,6 +1,3 @@
-################################################################
-# This test requires a physical robot
-################################################################
 import sys
 import numpy as np
 import math 
@@ -11,8 +8,12 @@ import time
 rootdir = os.path.dirname(os.path.dirname(__file__))
 os.sys.path.insert(0, rootdir)
 from roman import ur
-from roman.sim.ur import SimEnv 
+from roman.sim.ur_rq3 import SimEnv 
 
+#############################################################
+# Arm unit tests that bypass the manipulator server.
+# These tests can run on either sim or real arm.
+#############################################################
 def read_test(con):
     print(f"Running {__file__}::{read_test.__name__}()")
     con.connect()
@@ -39,12 +40,19 @@ def move_test(con):
     while not state.is_goal_reached():
         state = arm_ctrl(cmd)
 
+    cmd = ur.Command(target_position=ur.Joints(0, -math.pi/2, math.pi/2, -math.pi/2, -math.pi/2, 0))
+    state = arm_ctrl(cmd)
+    while not state.is_goal_reached():
+        state = arm_ctrl(cmd)
+
     print("Tool pose:" + str(state.tool_pose()))
     print("Joint positions:" + str(state.joint_positions()))
     con.disconnect()
     print("Passed.")    
 
-   
+#############################################################
+# Runner
+#############################################################
 def run(real_robot = False):
     if real_robot:
         #read_test(ur.Connection())
@@ -55,7 +63,6 @@ def run(real_robot = False):
         read_test(ur.SimConnection(env))
         move_test(ur.SimConnection(env))
         env.disconnect()
-   
-#env_test()
+
 if __name__ == '__main__':
-    run(True)
+    run(False)
