+++ src/test/arm_sim_tests.py
@@ -9,9 +9,12 @@ import pybullet as pb
 rootdir = os.path.dirname(os.path.dirname(__file__))
 os.sys.path.insert(0, rootdir)
 from roman.ur import *
-from roman.sim.ur import SimEnv
+from roman.sim.ur_rq3 import SimEnv
 from roman.ur.scripts.interface import *
 
+#############################################################
+# Unit tests that bypass the manipulation server 
+#############################################################
 def get_arm_state_test(env):
     print(f"Running {__file__}::{get_arm_state_test.__name__}()")
     state = State.fromarray(get_arm_state())
@@ -38,7 +41,7 @@ def move_arm_test(env):
     print(state.tool_pose())
     marker_visual_id = pb.createVisualShape(pb.GEOM_BOX, halfExtents=[0.005, 0.005, 0.005], rgbaColor=[1,0,0,1])
     #pb.createMultiBody(baseVisualShapeIndex=marker_visual_id, basePosition=state.tool_pose()[:3])
-    cmd = Command(target_position=Tool(-0.4, -0.4, 0.3,0, math.pi/2, math.pi))
+    cmd = Command(target_position=Tool(-0.4, -0.4, 0.3,0, math.pi, 0))
     pb.createMultiBody(baseVisualShapeIndex=marker_visual_id, basePosition=cmd.target_position()[:3])
     state = arm_ctrl(cmd)
     while not state.is_goal_reached():
@@ -52,11 +55,13 @@ def move_arm_test(env):
     
     print("Passed.")    
 
-   
+
+#############################################################
+# Runner
+#############################################################
 def run():
     env = SimEnv()
     env.reset()
-    env.update()
 
     get_arm_state_test(env)
     execute_arm_command_test(env)
@@ -64,6 +69,5 @@ def run():
 
     env.disconnect()
    
-#env_test()
 if __name__ == '__main__':
     run()
