+++ src/test/arm_upload_tests.py
@@ -11,8 +11,11 @@ print(rootdir)
 from roman import ur
 from roman.ur import loader
 
+#############################################################
+# These tests verify the UR script generation and uploading
+#############################################################
 def generate_script():
-    script = ur.Connection()._URConnection__generate_urscript()
+    script = ur.Connection()._Connection__generate_urscript()
     # outF = open("complete.script", "w")
     # outF.writelines(script)
     # outF.close()
@@ -23,20 +26,20 @@ def validate_script_syntax():
     Uploads all the scripts (by loading test.script) to validate their syntax.
     If the test hangs, edit test.script and remove the imports one by one (from bottom to top)
     '''
-    script_folder = os.path.join(os.path.join(rootdir, 'robot'), 'scripts')
-    defs = [f"UR_CLIENT_IP=\"{UR_DEFAULT_CLIENT_IP}\"", f"UR_CLIENT_PORT={UR_DEFAULT_CLIENT_PORT}"]
+    script_folder = os.path.join(os.path.join(os.path.join(rootdir, 'roman'), 'ur'), 'scripts')
+    defs = [f"UR_CLIENT_IP=\"{ur.UR_DEFAULT_CLIENT_IP}\"", f"UR_CLIENT_PORT={ur.UR_DEFAULT_CLIENT_PORT}"]
     rt_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    rt_socket.connect((UR_ROBOT_IP, UR_RT_PORT))
+    rt_socket.connect((ur.UR_ROBOT_IP, ur.UR_RT_PORT))
     print(f"loading test")
     script = loader.load_script(script_folder, "test", defs = defs)
     loader.socket_send_retry(rt_socket, script.encode('ascii'))
 
     reverse_conn = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    reverse_conn.bind((UR_DEFAULT_CLIENT_IP, UR_DEFAULT_CLIENT_PORT))
+    reverse_conn.bind((ur.UR_DEFAULT_CLIENT_IP, ur.UR_DEFAULT_CLIENT_PORT))
     reverse_conn.listen(1)
     print('Waiting for robot to confirm... ')
     ctrl_socket, addr = reverse_conn.accept() 
-    if (addr[0] != UR_ROBOT_IP):
+    if (addr[0] != ur.UR_ROBOT_IP):
         raise RuntimeError("Invalid client connection")
     print('Script uploaded.')
 
@@ -55,7 +58,9 @@ def test_script():
     loader.socket_send_retry(rt_socket, script.encode('ascii'))
     rt_socket.close()
 
-
+#############################################################
+# Runner
+#############################################################
 def run(real_robot = False):
     # run the tests
     generate_script()
@@ -64,6 +69,5 @@ def run(real_robot = False):
         test_script()
         validate_script_syntax()
    
-#env_test()
 if __name__ == '__main__':
     run()
